requirements.txt
langchain==0.1.11
langchain-google-genai==0.0.11
google-generativeai==0.3.0
streamlit==1.28.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
